pyproject.toml
[tool.poetry]
name = "data-discovery-agent"
version = "0.1.0"
description = "An ADK agent that profiles data in a dataset and dynamically creates context for the data"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "data_discovery_agent", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
pydantic = "^2.0.0"
google-cloud-bigquery = "^3.11.0"
google-cloud-storage = "^2.10.0"
google-cloud-discoveryengine = "^0.11.0"
google-cloud-datacatalog = "^3.11.0"
google-cloud-dataplex = "^1.6.0"
google-cloud-dlp = "^3.12.0"
google-cloud-datacatalog-lineage = "^0.4.0"
google-generativeai = "^0.8.0"
mcp = "^1.0.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.11.0"
pytest-cov = "^4.1.0"
freezegun = "^1.2.0"
responses = "^0.23.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

